backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
PyPDF2>=3.0.1
python-docx>=1.1.2
pymupdf>=1.24.0
